+++ src/app/features/forms/lp-form/thunks/getWalletBalanceThunk/index.ts
@@ -1,0 +1,21 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+
+import { RootState } from '../../../../../store';
+import { rejectThunkWithError } from '../../../../helpers/reject-thunk-with-error';
+
+export const getWalletBalanceThunk = createAsyncThunk<
+  Awaited<number | ReturnType<typeof rejectThunkWithError>>,
+  void,
+  { state: RootState }
+>('lpForm/getWalletBalance', async (_, thunkAPI) => {
+  try {
+    const amm = thunkAPI.getState().lpForm.amm;
+    if (!amm || !amm.signer) {
+      return 0;
+    }
+
+    return await amm.underlyingTokens();
+  } catch (err) {
+    return rejectThunkWithError(thunkAPI, err);
+  }
+});
